+++ examples/flutter_weather/test/app_test.dart
@@ -10,18 +10,11 @@ import 'package:weather_repository/weather_repository.dart';
 
 import 'helpers/hydrated_bloc.dart';
 
-class FakeColor extends Fake implements Color {}
-
 class MockThemeCubit extends MockCubit<Color> implements ThemeCubit {}
 
 class MockWeatherRepository extends Mock implements WeatherRepository {}
 
 void main() {
-  setUpAll(() {
-    initHydratedBloc();
-    registerFallbackValue(FakeColor());
-  });
-
   group('WeatherApp', () {
     late WeatherRepository weatherRepository;
 
@@ -30,7 +23,11 @@ void main() {
     });
 
     testWidgets('renders WeatherAppView', (tester) async {
-      await tester.pumpWidget(WeatherApp(weatherRepository: weatherRepository));
+      await mockHydratedStorage(() async {
+        await tester.pumpWidget(
+          WeatherApp(weatherRepository: weatherRepository),
+        );
+      });
       expect(find.byType(WeatherAppView), findsOneWidget);
     });
   });
@@ -46,24 +43,34 @@ void main() {
 
     testWidgets('renders WeatherPage', (tester) async {
       when(() => themeCubit.state).thenReturn(Colors.blue);
-      await tester.pumpWidget(
-        RepositoryProvider.value(
-          value: weatherRepository,
-          child: BlocProvider.value(value: themeCubit, child: WeatherAppView()),
-        ),
-      );
+      await mockHydratedStorage(() async {
+        await tester.pumpWidget(
+          RepositoryProvider.value(
+            value: weatherRepository,
+            child: BlocProvider.value(
+              value: themeCubit,
+              child: WeatherAppView(),
+            ),
+          ),
+        );
+      });
       expect(find.byType(WeatherPage), findsOneWidget);
     });
 
     testWidgets('has correct theme primary color', (tester) async {
       const color = Color(0xFFD2D2D2);
       when(() => themeCubit.state).thenReturn(color);
-      await tester.pumpWidget(
-        RepositoryProvider.value(
-          value: weatherRepository,
-          child: BlocProvider.value(value: themeCubit, child: WeatherAppView()),
-        ),
-      );
+      await mockHydratedStorage(() async {
+        await tester.pumpWidget(
+          RepositoryProvider.value(
+            value: weatherRepository,
+            child: BlocProvider.value(
+              value: themeCubit,
+              child: WeatherAppView(),
+            ),
+          ),
+        );
+      });
       final materialApp = tester.widget<MaterialApp>(find.byType(MaterialApp));
       expect(materialApp.theme?.primaryColor, color);
     });