+++ extensions/vscode/src/utils/analyze-dependencies.ts
@@ -1,8 +1,7 @@
 import * as _ from "lodash";
-
 import * as semver from "semver";
 import { window, env, Uri } from "vscode";
-import { getPubspec } from ".";
+import { getLatestPackageVersion, getPubspec } from ".";
 import { updatePubspecDependency } from "./update-pubspec-dependency";
 
 interface Dependency {
@@ -16,92 +15,60 @@ interface Action {
   callback: Function;
 }
 
+const openBlocMigrationGuide = {
+  name: "Open Migration Guide",
+  callback: () => {
+    env.openExternal(Uri.parse("https://bloclibrary.dev/#/migration"));
+  },
+};
+const openEquatableMigrationGuide = {
+  name: "Open Migration Guide",
+  callback: () => {
+    env.openExternal(
+      Uri.parse(
+        "https://github.com/felangel/equatable/blob/master/doc/migration_guides/migration-0.6.0.md"
+      )
+    );
+  },
+};
+
+const deps = [
+  { name: "angular_bloc", actions: [openBlocMigrationGuide] },
+  { name: "bloc", actions: [openBlocMigrationGuide] },
+  { name: "bloc_concurrency", actions: [openBlocMigrationGuide] },
+  { name: "equatable", actions: [openEquatableMigrationGuide] },
+  { name: "flutter_bloc", actions: [openBlocMigrationGuide] },
+  { name: "hydrated_bloc", actions: [openBlocMigrationGuide] },
+  { name: "replay_bloc", actions: [openBlocMigrationGuide] },
+  { name: "sealed_flutter_bloc", actions: [openBlocMigrationGuide] },
+];
+
+const devDeps = [{ name: "bloc_test", actions: [openBlocMigrationGuide] }];
+
 export async function analyzeDependencies() {
-  const openBlocMigrationGuide = {
-    name: "Open Migration Guide",
-    callback: () => {
-      env.openExternal(Uri.parse("https://bloclibrary.dev/#/migration"));
-    },
-  };
-  const dependenciesToAnalyze = [
-    {
-      name: "equatable",
-      version: "^2.0.3",
-      actions: [
-        {
-          name: "Open Migration Guide",
-          callback: () => {
-            env.openExternal(
-              Uri.parse(
-                "https://github.com/felangel/equatable/blob/master/doc/migration_guides/migration-0.6.0.md"
-              )
-            );
-          },
-        },
-      ],
-    },
-    {
-      name: "bloc",
-      version: "^7.2.1",
-      actions: [openBlocMigrationGuide],
-    },
-
-    {
-      name: "flutter_bloc",
-      version: "^7.3.3",
-      actions: [openBlocMigrationGuide],
-    },
-    {
-      name: "angular_bloc",
-      version: "^7.1.0",
-      actions: [openBlocMigrationGuide],
-    },
-    {
-      name: "hydrated_bloc",
-      version: "^7.1.0",
-      actions: [openBlocMigrationGuide],
-    },
-    {
-      name: "sealed_flutter_bloc",
-      version: "^7.1.0",
-      actions: [openBlocMigrationGuide],
-    },
-    {
-      name: "replay_bloc",
-      version: "^0.1.0",
-      actions: [openBlocMigrationGuide],
-    },
-    {
-      name: "bloc_concurrency",
-      version: "^0.1.0",
-      actions: [],
-    },
-  ];
-
-  const devDependenciesToAnalyze = [
-    {
-      name: "bloc_test",
-      version: "^8.5.0",
-      actions: [openBlocMigrationGuide],
-    },
-  ];
-
+  const dependencies = await getDependencies(deps);
+  const devDependencies = await getDependencies(devDeps);
   const pubspec = await getPubspec();
-  const dependencies = _.get(pubspec, "dependencies", {});
-  const devDependencies = _.get(pubspec, "dev_dependencies", {});
+  const pubspecDependencies = _.get(pubspec, "dependencies", {});
+  const pubspecDevDependencies = _.get(pubspec, "dev_dependencies", {});
 
-  checkForUpgrades(dependenciesToAnalyze, dependencies);
-  checkForUpgrades(devDependenciesToAnalyze, devDependencies);
+  checkForUpgrades(dependencies, pubspecDependencies);
+  checkForUpgrades(devDependencies, pubspecDevDependencies);
 }
 
 function checkForUpgrades(
-  dependenciesToAnalyze: Dependency[],
-  dependencies: object[]
+  dependencies: Dependency[],
+  pubspecDependencies: object[]
 ) {
-  for (let i = 0; i < dependenciesToAnalyze.length; i++) {
-    const dependency = dependenciesToAnalyze[i];
-    if (_.has(dependencies, dependency.name)) {
-      const dependencyVersion = _.get(dependencies, dependency.name, "latest");
+  for (let i = 0; i < dependencies.length; i++) {
+    const dependency = dependencies[i];
+    if (_.isEmpty(dependency.version)) continue;
+    if (_.has(pubspecDependencies, dependency.name)) {
+      const dependencyVersion = _.get(
+        pubspecDependencies,
+        dependency.name,
+        "latest"
+      );
       if (dependencyVersion === "latest") continue;
       if (dependencyVersion === "any") continue;
       if (dependencyVersion == null) continue;
@@ -121,7 +88,7 @@ function checkForUpgrades(
             if (invokedAction === "Update") {
               return updatePubspecDependency({
                 name: dependency.name,
-                latestVersion: dependency.version,
+                latestVersion: `^${dependency.version}`,
                 currentVersion: dependencyVersion,
               });
             }
@@ -135,4 +102,20 @@ function checkForUpgrades(
       }
     }
   }
+}
+
+async function getDependencies(
+  dependencies: { name: string; actions: Action[] }[]
+): Promise<Dependency[]> {
+  const futures: Promise<Dependency>[] = dependencies.map(
+    async (dependency) => {
+      return {
+        name: dependency.name,
+        actions: dependency.actions,
+        version: await getLatestPackageVersion(dependency.name),
+      };
+    }
+  );
+
+  return Promise.all(futures);
 }