+++ packages/bloc_test/test/legacy/cubits/exception_cubit.dart
@@ -1,7 +1,0 @@
-import 'package:bloc/bloc.dart';
-
-class ExceptionCubit extends Cubit<int> {
-  ExceptionCubit() : super(0);
-
-  void throwException(Exception e) => throw e;
-}