+++ packages/bloc_test/test/mock_bloc_test.dart
@@ -18,10 +18,6 @@ void main() {
   group('MockBloc', () {
     late CounterBloc counterBloc;
 
-    setUpAll(() {
-      registerFallbackValue(CounterEvent.increment);
-    });
-
     setUp(() {
       counterBloc = MockCounterBloc();
     });
@@ -32,13 +28,7 @@ void main() {
     });
 
     test('is compatible with listen', () {
-      // ignore: deprecated_member_use
-      expect(counterBloc.listen((_) {}), isA<StreamSubscription>());
       expect(counterBloc.stream.listen((_) {}), isA<StreamSubscription>());
-    });
-
-    test('is compatible with emit', () {
-      counterBloc.emit(10);
     });
 
     test('is compatible with add', () {
@@ -78,14 +68,6 @@ void main() {
       expect(counterBloc.close(), completes);
     });
 
-    test('is compatible with mapEventToState', () {
-      expect(
-        // ignore: deprecated_member_use
-        counterBloc.mapEventToState(CounterEvent.increment),
-        isA<Stream<int>>(),
-      );
-    });
-
     test('is automatically compatible with whenListen', () {
       whenListen(
         counterBloc,
@@ -97,19 +79,6 @@ void main() {
           <Matcher>[equals(0), equals(1), equals(2), equals(3), emitsDone],
         ),
       );
-    });
-
-    test('is automatically compatible with whenListen (legacy)', () {
-      final states = <int>[];
-      whenListen(
-        counterBloc,
-        Stream<int>.fromIterable([0, 1, 2, 3]),
-      );
-      // ignore: deprecated_member_use
-      counterBloc.listen(states.add, onDone: () {
-        expect(states, equals([0, 1, 2, 3]));
-        expect(counterBloc.state, equals(3));
-      });
     });
   });
 
@@ -136,19 +105,6 @@ void main() {
           <Matcher>[equals(0), equals(1), equals(2), equals(3), emitsDone],
         ),
       );
-    });
-
-    test('is automatically compatible with whenListen (legacy)', () {
-      final states = <int>[];
-      whenListen(
-        counterCubit,
-        Stream<int>.fromIterable([0, 1, 2, 3]),
-      );
-      // ignore: deprecated_member_use
-      counterCubit.listen(states.add, onDone: () {
-        expect(states, equals([0, 1, 2, 3]));
-        expect(counterCubit.state, equals(3));
-      });
     });
   });
 }